22.01.01

+++ califonia_house_price.py
@@ -3,13 +3,14 @@
 
 import os
 import tarfile
-import urllib
+import urllib.request
 
 DOWNLOAD_ROOT = "https://raw.githubusercontent.com/ageron/handson-ml2/master/"
 HOUSING_PATH = os.path.join("datasets", "housing")
 HOUSING_URL = DOWNLOAD_ROOT + "datasets/housing/housing.tgz"
 
 def fetch_housing_data(housing_url=HOUSING_URL, housing_path=HOUSING_PATH):
+    if os.path.isfile(os.path.join(housing_path, "housing.csv")): return
     os.makedirs(housing_path, exist_ok=True)
     tgz_path = os.path.join(housing_path, "housing.tgz")
     urllib.request.urlretrieve(housing_url, tgz_path)
@@ -18,12 +19,13 @@ def fetch_housing_data(housing_url=HOUSING_URL, housing_path=HOUSING_PATH):
     housing_tgz.close()
 
 
+
 import pandas as pd
 def load_housing_data(housing_path=HOUSING_PATH):
     csv_path = os.path.join(housing_path, "housing.csv")
     return pd.read_csv(csv_path)
 
-
+fetch_housing_data()
 housing = load_housing_data()
 print(housing.head(10))
 
@@ -37,7 +39,6 @@ print(housing["ocean_proximity"].value_counts())
 print(housing.describe())
 
 
-%matplotlib inline
 import matplotlib.pyplot as plt
 housing.hist(bins=50, figsize=(20,15))
 plt.show()
@@ -84,4 +85,5 @@ train_set, test_set = train_test_split(housing, test_size=0.2, random_state=42)
 
 
 housing["income_cat"] = pd.cut(housing["median_income"], bins=[0., 1.5, 3.0, 4.5, 6., np.inf], labels=[1, 2, 3, 4, 5])
-print(housing["income_cat"].hist())
+housing["income_cat"].hist()
+plt.show()
